moved filterData function from organisation index to orgHelpers
